Make more members private, which ought to fix any visibility problems.
There's something wrong if private members need to be exported, and it builds
fine for me now with visibility support enabled.

Comment out unused variables in obsolete functions.

+++ libcore/vm/ASHandlers.cpp
@@ -2859,7 +2859,7 @@ SWFHandlers::ActionCallFunction(ActionExec& thread)
 #endif
 
     std::auto_ptr< std::vector<as_value> > args ( new std::vector<as_value> );
-    for (int i=0; i<nargs; ++i)
+    for (size_t i=0; i<nargs; ++i)
         args->push_back(env.pop()); // TODO: drop !
 
     //log_debug("ActionCallFunction calling call_method with %p as this_ptr", this_ptr);
@@ -3577,7 +3577,7 @@ SWFHandlers::ActionCallMethod(ActionExec& thread)
 #endif
 
     std::auto_ptr< std::vector<as_value> > args ( new std::vector<as_value> );
-    for (int i=0; i<nargs; ++i)
+    for (size_t i=0; i<nargs; ++i)
         args->push_back(env.pop()); // TODO: drop !
 
     as_value result = call_method(method_val, &env, this_ptr,
